Moved is_openeye_installed to utils
